feat: implement ClockDisplay and DetailsPanel components with associated types, tests, and stories
- Created ClockDisplay component to show current time and greeting based on time of day.
- Added DetailsPanel component to display additional time-related information.
- Introduced QuoteSection component to fetch and display quotes with refresh functionality.
- Developed ToggleButton component for expanding/collapsing details.
- Implemented useQuote and useTime hooks for managing quote and time data.
- Added utility functions for time calculations and background image selection based on time of day.
- Created types for API responses and component props.
- Added tests and stories for all new components.
