+++ src/components/ClockDisplay/ClockDisplay.test.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import { render, screen } from '@testing-library/react'
+import ClockDisplay from './ClockDisplay'
+
+describe('ClockDisplay', () => {
+  it('should render the component', () => {
+    render(<ClockDisplay />)
+    expect(screen.getByTestId('clock-display')).toBeInTheDocument()
+  })
+})
