+++ src/components/ClockDisplay/ClockDisplay.types.ts
@@ -1,0 +1,8 @@
+import { ClockData, GreetingTime, TimeData } from '@/types/time.types'
+
+export interface ClockDisplayProps {
+  clockData: ClockData
+  timeData: TimeData
+  greetingTime: GreetingTime
+  isFooterExpanded: boolean
+}
