+++ src/components/DetailsPanel/DetailsPanel.types.ts
@@ -1,0 +1,6 @@
+import { TimeData } from '@/types/time.types'
+
+export interface DetailsPanelProps {
+  timeData: TimeData
+  expanded: boolean
+}
