+++ src/components/QuoteSection/QuoteSection.types.ts
@@ -1,0 +1,9 @@
+import { QuoteResponse } from '@/types/api.types'
+
+export interface QuoteSectionProps {
+  quote: QuoteResponse
+  loaded: boolean
+  refreshing: boolean
+  onRefresh: () => void
+  visible: boolean
+}
