+++ src/components/ToggleButton/ToggleButton.types.ts
@@ -1,0 +1,4 @@
+export interface ToggleButtonProps {
+  expanded: boolean
+  onToggle: () => void
+}
